probably the last version ig

+++ Final/Control_Flow_Features/Fan_In.cpp
@@ -1,3 +1,3 @@
-double FanIn(int numEdge,int numNode){
+double FanIn(double numEdge,int numNode){
     return (numEdge*1.0)/numNode;
 }
