+++ Final/Control_Flow_Features/Fan_Out.cpp
@@ -1,3 +1,3 @@
-double FanOut(int numEdge,int numNode){
+double FanOut(double numEdge,int numNode){
     return (numEdge*1.0)/numNode;
 }
